requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
numpy
python-dotenv
pdfplumber
pandas
pytest
